fix: improve FTA score sim-cloudfront-s3-origin
